+++ src/mapped-transition.js
@@ -1,0 +1,7 @@
+/** @returns {{ wasMapped: boolean, justBecameMapped: boolean }} */
+export function advanceMappedTracking(previousWasMapped, isMappedNow) {
+  return {
+    wasMapped: isMappedNow,
+    justBecameMapped: isMappedNow && !previousWasMapped,
+  };
+}